refactor: extract support components from AgentWorkspace
